@@ -1,6 +1,6 @@
 module AccessControl
   VERSION_MAJOR = 0
   VERSION_MINOR = 19
-  VERSION_TINY = 0
+  VERSION_TINY = 1
   VERSION = "#{VERSION_MAJOR}.#{VERSION_MINOR}.#{VERSION_TINY}"
 end
